Cleanup duplicate function to get port number from named port
Currently, the function to translate named port to port number is
located in two places (pod utils and endpointslice lib).
When fixing the bug in restartable init containers, one part of the code
was fixed, but the other part was not, leaving the bug unresolved.
To prevent such partial fixes in the future, we will make the function
in the endpointslice lib public and remove the other part of the code
from pod utils. Then consume the endpointslice lib in k/k.

Signed-off-by: Tsubasa Nagasawa <toversus2357@gmail.com>

Kubernetes-commit: 0ad351281bc203501ca3c40ebf9af2dae2e2c258