Minor updates to traffic distribution unit tests
Merge TestReconcileHints_trafficDistribution_is_PreferClose and
TestReconcileHints_trafficDistribution_is_nil_or_empty together.
Change the `trafficDistribution: ""` test to `trafficDistribution:
Unknown`, since `""` is not actually a possible value (but we should
still test that unknown values are ignored, to prevent weird skew
bugs).

Fill in the NodeName field in the endpoints. It's not needed yet but
it will be.

Kubernetes-commit: 413af836b381b763957b3bfe1c77070028915b5d